pyproject.toml
[tool.poetry]
name = "app restaurante"
version = "0.1.0"
description = ""
authors = ["Swarder0 <yago.matheus.ss0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
motor = "^3.7.1"
pydantic = {extras = ["email"], version = "^2.11.4"}
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "3.2.2"
python-multipart = "^0.0.20"
python-dotenv = "^1.1.0"
httpx = "^0.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
